requirements.txt
pytubefix
ffmpeg-python
openai